Global Functions Code Refactor
Created a Global Functions singleton and moved some functions from the game script over to there.

Current global functions has functions for adding sounds, both remotely and locally.

Working on adding functions for adding scenes.

Also decoupled the code for casting a decal and getting the hit result from the weapon raycast to allow for better separation in the future.